requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.2.0
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
babel==2.18.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
boto3==1.42.82
botocore==1.42.82
build==1.4.4
cachetools==5.5.2
catboost==1.2.10
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.45
dagshub-annotation-converter==0.2.0
databricks-sdk==0.102.0
dataclasses-json==0.6.7
datasets==4.8.5
ddgs==9.14.2
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
Django==5.2.14
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.18.0
dpath==2.2.0
dulwich==1.1.0
durationpy==0.10
dvc==3.67.1
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
ensure==1.0.4
entrypoints==0.4
et_xmlfile==2.0.0
eval_type_backport==0.3.1
executing==2.2.1
faiss-cpu==1.8.0
fastapi==0.135.3
fastembed==0.8.0
fastjsonschema==2.21.2
filelock==3.29.0
filetype==1.2.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.4.0
funcy==2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-run==0.16.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.67.0
google-generativeai==0.8.6
google-pasta==0.2.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
gql==4.0.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.5.0
groq==0.37.1
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.71.2
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huey==2.6.0
huggingface-hub==0.23.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.13.2
kiwisolver==1.4.9
kombu==5.6.2
kubernetes==35.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.27
langchain-google-genai==4.2.2
langchain-groq==0.2.2
langchain-mistralai==0.2.12
langchain-openai==0.1.25
langchain-protocol==0.0.15
langchain-text-splitters==0.3.4
langchainhub==0.1.21
langgraph==0.5.4
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.74
langsmith==0.1.147
lark==1.3.1
libclang==18.1.1
lightgbm==4.6.0
loguru==0.7.3
lxml==6.1.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistralai==2.0.0
mistune==3.2.0
ml-dtypes==0.3.2
mlflow==2.22.0
mlflow-skinny==2.22.0
mlflow-tracing==3.10.1
-e git+https://github.com/rahultomer218-bot/YT-MLOOPS-Project1.git@7089c5adf9511ee52696a24f0a2e4a38cac359af#egg=mlops_project
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
notebook==7.5.4
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
odfpy==1.4.1
omegaconf==2.3.0
onnxruntime==1.23.2
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.9
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pathvalidate==3.3.1
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==11.3.0
pip_system_certs==5.3
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
posthog==7.13.1
prettytable==3.17.0
primp==1.2.3
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic-settings==2.14.0
pydantic_core==2.46.4
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.2
Pygments==2.20.0
pygtrie==2.5.0
pymongo==4.17.0
PyMySQL==1.1.3
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.3
python-box==7.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-json-logger==4.0.0
python-multipart==0.0.27
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==15.0.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.2
seaborn==0.13.2
semver==3.0.4
Send2Trash==2.1.0
sentence-transformers==5.4.1
sentencepiece==0.2.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.35
sqlparse==0.5.5
sqltrie==0.11.2
stack-data==0.6.3
starlette==1.0.0
streamlit==1.57.0
sympy==1.14.0
tabulate==0.10.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
terminado==0.18.1
tf_keras==2.16.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.42.4
treelib==1.8.0
typer==0.25.1
types-requests==2.33.0.20260508
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.7.0
-e git+https://github.com/rahultomer218-bot/US-Visa-MLops-Learning-Machine-Language.git@89d8b3b0ad63c34c7142695995d55222e4b04a9d#egg=US_Visa_MLops_Learning_Machine_Language
uuid_utils==0.14.1
uv==0.10.10
uvicorn==0.42.0
uvloop==0.22.1
vine==5.1.0
voluptuous==0.16.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
wordcloud==1.9.6
wrapt==1.17.3
xgboost==3.2.0
xxhash==3.7.0
yarl==1.23.0
youtube-transcript-api==1.2.4
zc.lockfile==4.0
zipp==3.23.0
zstandard==0.25.0
